requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
joblib
jinja2
python-multipart

langchain
langchain-ollama
langchain-chroma
PyPDF2
chromadb

passlib[bcrypt]
python-jose[cryptography]
python-dotenv
